Extract `tecdsa.PrivateKeyShare` marshaling code to a separate file
